requirements.txt
lxml==4.4.1
python-slugify==3.0.4
requests==2.22.0
